+++ packages/orq-rc/src/orq_ai_sdk/models/createbudgetop.py
@@ -211,5 +211,5 @@ class CreateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-03T10:17:05.785Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-03T11:43:27.908Z")
     r"""The date and time the resource was last updated"""