+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -119,5 +119,5 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-03T10:17:05.785Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-03T11:43:27.908Z")
     r"""The date and time the resource was last updated"""